Added JXS specific header files to separate JXS functionality out. Updated autoconf to switch JXS support for ADCP on or off.
