+++ src/AS_DCP.h
@@ -1456,229 +1456,6 @@ namespace ASDCP {
 
   //---------------------------------------------------------------------------------
   //
-  namespace JXS 
-  {
-	  const ui32_t MaxComponents = 4; // ISO 21122-1 Annex A.2 up to 8 components
-	  const ui32_t MaxHorizontalLevels = 15;
-	  const ui32_t MaxVerticalLevels = 2;
-
-#pragma pack(1)
-	  struct ImageComponent_t  // Essentially, a lookalike of the CDT marker, just with less bit-packing
-	  {
-		  ui8_t Bc;  // Bitdepth (literal, not -1)
-		  ui8_t Sx;
-		  ui8_t Sy;  // Subsampling factors, horizontal and vertically. Bit-packed in the marker.
-	  };
-#pragma pack()
-
-	  struct PictureDescriptor
-	  {
-		  Rational       EditRate;
-		  ui32_t         ContainerDuration;
-		  Rational       SampleRate;
-		  ui32_t         StoredWidth;
-		  ui32_t         StoredHeight;
-		  Rational       AspectRatio;
-		  ui16_t         Ppih; // Profile, copy from the PIH marker
-		  ui16_t         Plev; // Level and sublevel, copy from the PIH marker
-		  ui16_t         Wf;  // Frame width, copy from the PIH marker
-		  ui16_t         Hf;  // Frame height, copy from the PIH marker
-		  ui16_t         Hsl; // slice height, copy from the PIH marker
-		  ui16_t         Cw;  // column width, or 0 for no columns, copy from the PIH marker
-		  ui8_t          Nc;  // number of components, copy from the PIH marker
-	          ui32_t         MaximumBitRate; // bit rate in MB/sec, or 0 if not known
-	          ui8_t          Primaries; // Color primaries as defined by CICP
-	          ui8_t          TransferCurve; // Transfer curve as defined by CICP
-	          ui8_t          Matrix; // Transform matrix, as defined by CICP
-	          bool           fullRange; // If true, no head and toe region
-		  ImageComponent_t ImageComponents[MaxComponents]; // These are copies from the CDT (component table)
-	  };
-	  // Print debugging information to std::ostream
-	  std::ostream& operator << (std::ostream& strm, const PictureDescriptor& pdesc);
-	  // Print debugging information to stream (stderr default)
-	  void   PictureDescriptorDump(const PictureDescriptor&, FILE* = 0);
-
-	  //
-	  class FrameBuffer : public ASDCP::FrameBuffer
-	  {
-	  public:
-		  FrameBuffer() {}
-		  FrameBuffer(ui32_t size) { Capacity(size); }
-		  virtual ~FrameBuffer() {}
-
-		  // Print debugging information to stream (stderr default)
-		  void Dump(FILE* = 0, ui32_t dump_bytes = 0) const;
-	  };
-
-
-	  // An object which opens and reads a JPEG XS codestream file.  The file is expected
-	  // to contain exactly one complete frame of picture essence as an unwrapped (raw)
-	  // ISO/IEC 21122 codestream.
-	  class CodestreamParser
-	  {
-		  class h__CodestreamParser;
-		  mem_ptr<h__CodestreamParser> m_Parser;
-		  ASDCP_NO_COPY_CONSTRUCT(CodestreamParser);
-
-	  public:
-		  CodestreamParser();
-		  virtual ~CodestreamParser();
-
-		  // Opens a file for reading, parses enough data to provide a complete
-			  // set of stream metadata for the MXFWriter below.
-		  // The frame buffer's PlaintextOffset parameter will be set to the first
-		  // byte of the data segment. Set this value to zero if you want
-		  // encrypted headers.
-		  Result_t OpenReadFrame(const std::string& filename, FrameBuffer&) const;
-
-		  // Fill a PictureDescriptor struct with the values from the file's codestream.
-		  // Returns RESULT_INIT if the file is not open.
-		  Result_t FillPictureDescriptor(PictureDescriptor&) const;
-	  };
-
-	  // Parses the data in the frame buffer to fill in the picture descriptor. Copies
-	  // the offset of the image data into start_of_data. Returns error if the parser fails.
-	  Result_t ParseMetadataIntoDesc(const FrameBuffer&, PictureDescriptor&, byte_t* start_of_data = 0);
-
-	  // An object which reads a sequence of files containing JPEG XS pictures.
-	  class SequenceParser
-	  {
-		  class h__SequenceParser;
-		  mem_ptr<h__SequenceParser> m_Parser;
-		  ASDCP_NO_COPY_CONSTRUCT(SequenceParser);
-
-	  public:
-		  SequenceParser();
-		  virtual ~SequenceParser();
-
-		  // Opens a directory for reading.  The directory is expected to contain one or
-		  // more files, each containing the codestream for exactly one picture. The
-		  // files must be named such that the frames are in temporal order when sorted
-		  // alphabetically by filename. The parser will automatically parse enough data
-		  // from the first file to provide a complete set of stream metadata for the
-		  // MXFWriter below.  If the "pedantic" parameter is given and is true, the
-		  // parser will check the metadata for each codestream and fail if a
-		  // mismatch is detected.
-		  Result_t OpenRead(const std::string& filename) const;
-
-		  // Opens a file sequence for reading.  The sequence is expected to contain one or
-		  // more filenames, each naming a file containing the codestream for exactly one
-		  // picture. The parser will automatically parse enough data
-		  // from the first file to provide a complete set of stream metadata for the
-		  // MXFWriter below.  If the "pedantic" parameter is given and is true, the
-		  // parser will check the metadata for each codestream and fail if a
-		  // mismatch is detected.
-		  Result_t OpenRead(const std::list<std::string>& file_list) const;
-
-		  // Fill a PictureDescriptor struct with the values from the first file's codestream.
-		  // Returns RESULT_INIT if the directory is not open.
-		  Result_t FillPictureDescriptor(PictureDescriptor&) const;
-
-		  // Rewind the directory to the beginning.
-		  Result_t Reset() const;
-
-		  // Reads the next sequential frame in the directory and places it in the
-		  // frame buffer. Fails if the buffer is too small or the direcdtory
-		  // contains no more files.
-		  // The frame buffer's PlaintextOffset parameter will be set to the first
-		  // byte of the data segment. Set this value to zero if you want
-		  // encrypted headers.
-		  Result_t ReadFrame(FrameBuffer&) const;
-	  };
-
-
-	  //
-	  class MXFWriter
-	  {
-		  class h__Writer;
-		  mem_ptr<h__Writer> m_Writer;
-		  ASDCP_NO_COPY_CONSTRUCT(MXFWriter);
-
-	  public:
-		  MXFWriter();
-		  virtual ~MXFWriter();
-
-		  // Warning: direct manipulation of MXF structures can interfere
-		  // with the normal operation of the wrapper.  Caveat emptor!
-		  virtual MXF::OP1aHeader& OP1aHeader();
-		  virtual MXF::OPAtomIndexFooter& OPAtomIndexFooter();
-		  virtual MXF::RIP& RIP();
-
-		  // Open the file for writing. The file must not exist. Returns error if
-		  // the operation cannot be completed or if nonsensical data is discovered
-		  // in the essence descriptor.
-		  Result_t OpenWrite(const std::string& filename, const WriterInfo&,
-			  const PictureDescriptor&, ui32_t HeaderSize = 16384);
-
-		  // Writes a frame of essence to the MXF file. If the optional AESEncContext
-		  // argument is present, the essence is encrypted prior to writing.
-		  // Fails if the file is not open, is finalized, or an operating system
-		  // error occurs.
-		  Result_t WriteFrame(const FrameBuffer&, AESEncContext* = 0, HMACContext* = 0);
-
-		  // Closes the MXF file, writing the index and revised header.
-		  Result_t Finalize();
-	  };
-
-	  //
-	  class MXFReader
-	  {
-		  class h__Reader;
-		  mem_ptr<h__Reader> m_Reader;
-		  ASDCP_NO_COPY_CONSTRUCT(MXFReader);
-
-	  public:
-		  MXFReader();
-		  virtual ~MXFReader();
-
-		  // Warning: direct manipulation of MXF structures can interfere
-		  // with the normal operation of the wrapper.  Caveat emptor!
-		  virtual MXF::OP1aHeader& OP1aHeader();
-		  virtual MXF::OPAtomIndexFooter& OPAtomIndexFooter();
-		  virtual MXF::RIP& RIP();
-
-		  // Open the file for reading. The file must exist. Returns error if the
-		  // operation cannot be completed.
-		  Result_t OpenRead(const std::string& filename) const;
-
-		  // Returns RESULT_INIT if the file is not open.
-		  Result_t Close() const;
-
-		  // Fill an AudioDescriptor struct with the values from the file's header.
-		  // Returns RESULT_INIT if the file is not open.
-		  Result_t FillPictureDescriptor(PictureDescriptor&) const;
-
-		  // Fill a WriterInfo struct with the values from the file's header.
-		  // Returns RESULT_INIT if the file is not open.
-		  Result_t FillWriterInfo(WriterInfo&) const;
-
-		  // Compute the necessary frame buffer size in bytes for CBR
-		  // codestreamss.
-		  Result_t CalcFrameBufferSize(ui64_t &size) const;
-	  
-		  // Reads a frame of essence from the MXF file. If the optional AESEncContext
-		  // argument is present, the essence is decrypted after reading. If the MXF
-		  // file is encrypted and the AESDecContext argument is NULL, the frame buffer
-		  // will contain the ciphertext frame data. If the HMACContext argument is
-		  // not NULL, the HMAC will be calculated (if the file supports it).
-		  // Returns RESULT_INIT if the file is not open, failure if the frame number is
-		  // out of range, or if optional decrypt or HAMC operations fail.
-		  Result_t ReadFrame(ui32_t frame_number, FrameBuffer&, AESDecContext* = 0, HMACContext* = 0) const;
-
-		  // Using the index table read from the footer partition, lookup the frame number
-		  // and return the offset into the file at which to read that frame of essence.
-		  // Returns RESULT_INIT if the file is not open, and RESULT_FRAME if the frame number is
-		  // out of range.
-		  Result_t LocateFrame(ui32_t FrameNum, Kumu::fpos_t& streamOffset, i8_t& temporalOffset, i8_t& keyFrameOffset) const;
-
-		  // Print debugging information to stream
-		  void     DumpHeaderMetadata(FILE* = 0) const;
-		  void     DumpIndex(FILE* = 0) const;
-	  };
-  }
-
-  //---------------------------------------------------------------------------------
-  //
   namespace TimedText
     {
       enum MIMEType_t { MT_BIN, MT_PNG, MT_OPENTYPE };
