+++ src/AS_DCP_JXS.cpp
@@ -31,6 +31,7 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
 #include "AS_DCP_internal.h"
+#include "AS_DCP_JXS.h"
 #include <iostream>
 #include <iomanip>
 #include <algorithm>
